removed /team and combined in /about
